complete the implementation of remove function
